+++ rs/quantization/src/lib.rs
@@ -1,5 +1,6 @@
 #![feature(portable_simd)]
 pub mod pq;
 pub mod pq_builder;
+pub mod rabitq;
 pub mod rabitq_builder;
 pub mod quantization;
